Decided to stop working on the v-usb solution.
After tens of hour it became clear that basing the project on running
v-usb and nrflite on the atmega328p chip is impossible or at least
really difficult. I was able to make the v-usb work and also able to
make the nrf24l01 work, but I wasn't able to make it work together. One
or the other library experienced timing issues and wasn't working
properly. I tried to slightly change the nrflite library but it didn't
help. I do not have time to spend any more time on this path, so I
decided to try different aproach. Next commit will update the files to
state when I decided to quit.